+++ CHANGELOG.md
@@ -12,6 +12,7 @@ for more details on upgrading to v4.
   - Remove elasticsearch integrations and replace with redisearch
   - Refactor all query endpoints, including changing query inputs (Breaking Change)
   - Changed route for `GET /v1/contract` from a query to a list (Breaking Change)
+  - Add support to disable schedule when updating a smart contract
 - **Documentation:**
   - Update documentation for redisearch
   - Add new page for migration considerations when upgrading from v3 to v4